% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_funs.R
\name{package_stats}
\alias{package_stats}
\title{Provides package statistics}
\usage{
package_stats()
}
\value{
Doesn't return anything, just a prompt in the console.
}
\description{
Summary statistics of a packages: number of lines, number of functions, etc...
}
\details{
This function looks for files in the \code{R/} and \code{src/} folders and gives some stats. If there is no \code{R/} folder directly accessible from the working directory, there will be no stats displayed.

Why this function? Well, it's just some goodies for package developers trying to be user-friendly!

The number of documentation lines (and number of words) corresponds to the number of non-empty roxygen documentation lines. So if you don't document your code with roxygen, well, this stat won't prompt.

Code lines correspond to non-commented, non-empty lines (by non empty: at least one letter must appear).

Comment lines are non-empty comments.
}
\examples{

package_stats()

}
