% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_funs.R
\name{fit_screen}
\alias{fit_screen}
\title{Nicely fits a message in the current R console}
\usage{
fit_screen(msg, width = 0.9)
}
\arguments{
\item{msg}{Text message: character vector.}

\item{width}{The maximum width of the screen the message should take. Default is 0.9.}
}
\value{
It returns a single character vector with line breaks at the appropriate width.
}
\description{
Utility to display long messages with nice formatting. This function cuts the message to fit the current screen width of the R console. Words are never cut in the middle.
}
\details{
This function does not handle tabulations.
}
\examples{

cat(fit_screen(enumerate_items(state.name, nmax = Inf)))

cat(fit_screen(enumerate_items(state.name, nmax = Inf), 0.5))

}
